Got the Triangle's toggle button working to switch modes
* Mode changes are also broadcast over all Sockets

+++ Libraries/TriangleLibrary/TriangleLights.h
@@ -21,7 +21,7 @@ extern Triangle *triangles;
 #define PUSH_BUTTON_INTERRUPT 1
 #define PUSH_BUTTON_PIN       3
 
-int get_button_value();
+byte get_button_value();
 
 /* Photo sensor */
 #define PHOTO_PIN A0
